Split `KeyBindingRow` out into partial pieces
